♻️Refactor: 검색창(SearchBar) 컴포넌트로 독립, 기존 코드 수정
